Extract Envelope::Item to its own file
